refactor(api): split Stripe dispute handlers out of webhooks.rs (#444)
webhooks.rs 2504 -> 1277 lines (-1227). Extract the Stripe
charge.dispute.* cluster (2 types, 5 helpers, 4 handlers, 10
DB-coupled e2e tests) into a new webhooks_disputes.rs. The four
handlers become pub(super) so the sibling webhooks module's
stripe_webhook still calls them via a `use super::webhooks_disputes`
import; the three route-registered #[handler] fns stay in webhooks.rs.

Path-B split (no #[OpenApi] type involved): create_combined_api is
unchanged, so no combined-API tuple slot is consumed and the OpenAPI
spec is unaffected. Verified byte-identical: a new committed guard
openapi::spec_snapshot renders create_combined_api to canonical JSON
and asserts a stable hash + counts (187 paths, 327 schemas, hash
4549fcf2...). Before/after `diff` of the dumped spec is empty
(472132 bytes both).

- cargo clippy -p api --tests --all-targets -> 0 warnings, 0 errors
- cargo nextest (webhooks + webhooks_disputes + spec_snapshot) -> 30/30 pass
